Move PhysicsServer largely out of main.cpp
Moves PhysicsServer largely out of main.cpp.
