+++ main/main.cpp
@@ -93,7 +93,7 @@
 
 #ifndef PHYSICS_2D_DISABLED
 #include "servers/physics_2d/physics_server_2d.h"
-#include "servers/physics_2d/physics_server_2d_dummy.h"
+#include "servers/physics_2d/physics_server_2d_manager.h"
 #endif // PHYSICS_2D_DISABLED
 
 // 3D
@@ -103,7 +103,7 @@
 
 #ifndef PHYSICS_3D_DISABLED
 #include "servers/physics_3d/physics_server_3d.h"
-#include "servers/physics_3d/physics_server_3d_dummy.h"
+#include "servers/physics_3d/physics_server_3d_manager.h"
 #endif // PHYSICS_3D_DISABLED
 
 #ifndef XR_DISABLED
@@ -185,14 +185,6 @@ static DisplayServer *display_server = nullptr;
 static RenderingServer *rendering_server = nullptr;
 static TextServerManager *tsman = nullptr;
 static ThemeDB *theme_db = nullptr;
-#ifndef PHYSICS_2D_DISABLED
-static PhysicsServer2DManager *physics_server_2d_manager = nullptr;
-static PhysicsServer2D *physics_server_2d = nullptr;
-#endif // PHYSICS_2D_DISABLED
-#ifndef PHYSICS_3D_DISABLED
-static PhysicsServer3DManager *physics_server_3d_manager = nullptr;
-static PhysicsServer3D *physics_server_3d = nullptr;
-#endif // PHYSICS_3D_DISABLED
 #ifndef XR_DISABLED
 static XRServer *xr_server = nullptr;
 #endif // XR_DISABLED
@@ -354,61 +346,6 @@ static Vector<String> get_files_with_extension(const String &p_root, const Strin
 	return paths;
 }
 #endif
-
-// FIXME: Could maybe be moved to have less code in main.cpp.
-void initialize_physics() {
-#ifndef PHYSICS_3D_DISABLED
-	/// 3D Physics Server
-	physics_server_3d = PhysicsServer3DManager::get_singleton()->new_server(
-			GLOBAL_GET(PhysicsServer3DManager::setting_property_name));
-	if (!physics_server_3d) {
-		// Physics server not found, Use the default physics
-		physics_server_3d = PhysicsServer3DManager::get_singleton()->new_default_server();
-	}
-
-	// Fall back to dummy if no default server has been registered.
-	if (!physics_server_3d) {
-		WARN_PRINT(vformat("Falling back to dummy PhysicsServer3D; 3D physics functionality will be disabled. If this is intended, set the %s project setting to Dummy.", PhysicsServer3DManager::setting_property_name));
-		physics_server_3d = memnew(PhysicsServer3DDummy);
-	}
-
-	// Should be impossible, but make sure it's not null.
-	ERR_FAIL_NULL_MSG(physics_server_3d, "Failed to initialize PhysicsServer3D.");
-	physics_server_3d->init();
-#endif // PHYSICS_3D_DISABLED
-
-#ifndef PHYSICS_2D_DISABLED
-	// 2D Physics server
-	physics_server_2d = PhysicsServer2DManager::get_singleton()->new_server(
-			GLOBAL_GET(PhysicsServer2DManager::get_singleton()->setting_property_name));
-	if (!physics_server_2d) {
-		// Physics server not found, Use the default physics
-		physics_server_2d = PhysicsServer2DManager::get_singleton()->new_default_server();
-	}
-
-	// Fall back to dummy if no default server has been registered.
-	if (!physics_server_2d) {
-		WARN_PRINT(vformat("Falling back to dummy PhysicsServer2D; 2D physics functionality will be disabled. If this is intended, set the %s project setting to Dummy.", PhysicsServer2DManager::setting_property_name));
-		physics_server_2d = memnew(PhysicsServer2DDummy);
-	}
-
-	// Should be impossible, but make sure it's not null.
-	ERR_FAIL_NULL_MSG(physics_server_2d, "Failed to initialize PhysicsServer2D.");
-	physics_server_2d->init();
-#endif // PHYSICS_2D_DISABLED
-}
-
-void finalize_physics() {
-#ifndef PHYSICS_3D_DISABLED
-	physics_server_3d->finish();
-	memdelete(physics_server_3d);
-#endif // PHYSICS_3D_DISABLED
-
-#ifndef PHYSICS_2D_DISABLED
-	physics_server_2d->finish();
-	memdelete(physics_server_2d);
-#endif // PHYSICS_2D_DISABLED
-}
 
 void finalize_display() {
 	rendering_server->finish();
@@ -773,10 +710,10 @@ Error Main::test_setup() {
 	}
 
 #ifndef PHYSICS_3D_DISABLED
-	physics_server_3d_manager = memnew(PhysicsServer3DManager);
+	PhysicsServer3DManager::initialize_server_manager();
 #endif // PHYSICS_3D_DISABLED
 #ifndef PHYSICS_2D_DISABLED
-	physics_server_2d_manager = memnew(PhysicsServer2DManager);
+	PhysicsServer2DManager::initialize_server_manager();
 #endif // PHYSICS_2D_DISABLED
 
 #ifndef NAVIGATION_2D_DISABLED
@@ -944,10 +881,10 @@ void Main::test_cleanup() {
 	memdelete(translation_server);
 	memdelete(tsman);
 #ifndef PHYSICS_3D_DISABLED
-	memdelete(physics_server_3d_manager);
+	PhysicsServer3DManager::finalize_server_manager();
 #endif // PHYSICS_3D_DISABLED
 #ifndef PHYSICS_2D_DISABLED
-	memdelete(physics_server_2d_manager);
+	PhysicsServer2DManager::finalize_server_manager();
 #endif // PHYSICS_2D_DISABLED
 	memdelete(globals);
 
@@ -3179,10 +3116,10 @@ Error Main::setup2(bool p_show_boot_logo) {
 	}
 
 #ifndef PHYSICS_3D_DISABLED
-	physics_server_3d_manager = memnew(PhysicsServer3DManager);
+	PhysicsServer3DManager::initialize_server_manager();
 #endif // PHYSICS_3D_DISABLED
 #ifndef PHYSICS_2D_DISABLED
-	physics_server_2d_manager = memnew(PhysicsServer2DManager);
+	PhysicsServer2DManager::initialize_server_manager();
 #endif // PHYSICS_2D_DISABLED
 
 #ifndef NAVIGATION_2D_DISABLED
@@ -3371,10 +3308,10 @@ Error Main::setup2(bool p_show_boot_logo) {
 			memdelete(input);
 			memdelete(tsman);
 #ifndef PHYSICS_3D_DISABLED
-			memdelete(physics_server_3d_manager);
+			PhysicsServer3DManager::finalize_server_manager();
 #endif // PHYSICS_3D_DISABLED
 #ifndef PHYSICS_2D_DISABLED
-			memdelete(physics_server_2d_manager);
+			PhysicsServer2DManager::finalize_server_manager();
 #endif // PHYSICS_2D_DISABLED
 
 			return err;
@@ -3822,7 +3759,12 @@ Error Main::setup2(bool p_show_boot_logo) {
 
 	MAIN_PRINT("Main: Load Physics");
 
-	initialize_physics();
+#ifndef PHYSICS_2D_DISABLED
+	PhysicsServer2DManager::initialize_server();
+#endif // PHYSICS_2D_DISABLED
+#ifndef PHYSICS_3D_DISABLED
+	PhysicsServer3DManager::initialize_server();
+#endif // PHYSICS_3D_DISABLED
 
 	register_server_singletons();
 
@@ -5269,7 +5211,13 @@ void Main::cleanup(bool p_force) {
 	NavigationServer3DManager::finalize_server();
 	NavigationServer3DManager::finalize_server_manager();
 #endif // NAVIGATION_3D_DISABLED
-	finalize_physics();
+
+#ifndef PHYSICS_2D_DISABLED
+	PhysicsServer2DManager::finalize_server();
+#endif // PHYSICS_2D_DISABLED
+#ifndef PHYSICS_3D_DISABLED
+	PhysicsServer3DManager::finalize_server();
+#endif // PHYSICS_3D_DISABLED
 
 	GDExtensionManager::get_singleton()->deinitialize_extensions(GDExtension::INITIALIZATION_LEVEL_SERVERS);
 	uninitialize_modules(MODULE_INITIALIZATION_LEVEL_SERVERS);
@@ -5299,12 +5247,12 @@ void Main::cleanup(bool p_force) {
 	memdelete(input_map);
 	memdelete(translation_server);
 	memdelete(tsman);
-#ifndef PHYSICS_3D_DISABLED
-	memdelete(physics_server_3d_manager);
-#endif // PHYSICS_3D_DISABLED
 #ifndef PHYSICS_2D_DISABLED
-	memdelete(physics_server_2d_manager);
+	PhysicsServer2DManager::finalize_server_manager();
 #endif // PHYSICS_2D_DISABLED
+#ifndef PHYSICS_3D_DISABLED
+	PhysicsServer3DManager::finalize_server_manager();
+#endif // PHYSICS_3D_DISABLED
 	memdelete(globals);
 
 	if (OS::get_singleton()->is_restart_on_exit_set()) {
