+++ modules/godot_physics_2d/register_types.cpp
@@ -35,6 +35,7 @@
 #include "core/config/project_settings.h"
 #include "core/object/callable_mp.h"
 #include "servers/physics_2d/physics_server_2d.h"
+#include "servers/physics_2d/physics_server_2d_manager.h"
 #include "servers/physics_2d/physics_server_2d_wrap_mt.h"
 
 static PhysicsServer2D *_createGodotPhysics2DCallback() {
