+++ servers/physics_2d/physics_server_2d.cpp
@@ -349,91 +349,3 @@ PhysicsServer2D::PhysicsServer2D() {
 PhysicsServer2D::~PhysicsServer2D() {
 	singleton = nullptr;
 }
-
-PhysicsServer2DManager *PhysicsServer2DManager::singleton = nullptr;
-const String PhysicsServer2DManager::setting_property_name(PNAME("physics/2d/physics_engine"));
-
-void PhysicsServer2DManager::on_servers_changed() {
-	String physics_servers("DEFAULT");
-	for (int i = get_servers_count() - 1; 0 <= i; --i) {
-		physics_servers += "," + get_server_name(i);
-	}
-	ProjectSettings::get_singleton()->set_custom_property_info(PropertyInfo(Variant::STRING, setting_property_name, PROPERTY_HINT_ENUM, physics_servers));
-	ProjectSettings::get_singleton()->set_restart_if_changed(setting_property_name, true);
-	ProjectSettings::get_singleton()->set_as_basic(setting_property_name, true);
-}
-
-void PhysicsServer2DManager::_bind_methods() {
-	ClassDB::bind_method(D_METHOD("register_server", "name", "create_callback"), &PhysicsServer2DManager::register_server);
-	ClassDB::bind_method(D_METHOD("set_default_server", "name", "priority"), &PhysicsServer2DManager::set_default_server);
-}
-
-PhysicsServer2DManager *PhysicsServer2DManager::get_singleton() {
-	return singleton;
-}
-
-void PhysicsServer2DManager::register_server(const String &p_name, const Callable &p_create_callback) {
-	//ERR_FAIL_COND(!p_create_callback.is_valid());
-	ERR_FAIL_COND(find_server_id(p_name) != -1);
-	physics_2d_servers.push_back(ClassInfo(p_name, p_create_callback));
-	on_servers_changed();
-}
-
-void PhysicsServer2DManager::set_default_server(const String &p_name, int p_priority) {
-	const int id = find_server_id(p_name);
-	ERR_FAIL_COND(id == -1); // Not found
-	if (default_server_priority < p_priority) {
-		default_server_id = id;
-		default_server_priority = p_priority;
-	}
-}
-
-int PhysicsServer2DManager::find_server_id(const String &p_name) {
-	for (int i = physics_2d_servers.size() - 1; 0 <= i; --i) {
-		if (p_name == physics_2d_servers[i].name) {
-			return i;
-		}
-	}
-	return -1;
-}
-
-int PhysicsServer2DManager::get_servers_count() {
-	return physics_2d_servers.size();
-}
-
-String PhysicsServer2DManager::get_server_name(int p_id) {
-	ERR_FAIL_INDEX_V(p_id, get_servers_count(), "");
-	return physics_2d_servers[p_id].name;
-}
-
-PhysicsServer2D *PhysicsServer2DManager::new_default_server() {
-	if (default_server_id == -1) {
-		return nullptr;
-	}
-	Variant ret;
-	Callable::CallError ce;
-	physics_2d_servers[default_server_id].create_callback.callp(nullptr, 0, ret, ce);
-	ERR_FAIL_COND_V(ce.error != Callable::CallError::CALL_OK, nullptr);
-	return Object::cast_to<PhysicsServer2D>(ret.get_validated_object());
-}
-
-PhysicsServer2D *PhysicsServer2DManager::new_server(const String &p_name) {
-	int id = find_server_id(p_name);
-	if (id == -1) {
-		return nullptr;
-	} else {
-		Variant ret;
-		Callable::CallError ce;
-		physics_2d_servers[id].create_callback.callp(nullptr, 0, ret, ce);
-		ERR_FAIL_COND_V(ce.error != Callable::CallError::CALL_OK, nullptr);
-		return Object::cast_to<PhysicsServer2D>(ret.get_validated_object());
-	}
-}
-
-PhysicsServer2DManager::PhysicsServer2DManager() {
-	singleton = this;
-}
-
-PhysicsServer2DManager::~PhysicsServer2DManager() {
-	singleton = nullptr;
-}
