+++ servers/physics_3d/physics_server_3d.cpp
@@ -549,91 +549,3 @@ PhysicsServer3D::PhysicsServer3D() {
 PhysicsServer3D::~PhysicsServer3D() {
 	singleton = nullptr;
 }
-
-PhysicsServer3DManager *PhysicsServer3DManager::singleton = nullptr;
-const String PhysicsServer3DManager::setting_property_name(PNAME("physics/3d/physics_engine"));
-
-void PhysicsServer3DManager::on_servers_changed() {
-	String physics_servers2("DEFAULT");
-	for (int i = get_servers_count() - 1; 0 <= i; --i) {
-		physics_servers2 += "," + get_server_name(i);
-	}
-	ProjectSettings::get_singleton()->set_custom_property_info(PropertyInfo(Variant::STRING, setting_property_name, PROPERTY_HINT_ENUM, physics_servers2));
-	ProjectSettings::get_singleton()->set_restart_if_changed(setting_property_name, true);
-	ProjectSettings::get_singleton()->set_as_basic(setting_property_name, true);
-}
-
-void PhysicsServer3DManager::_bind_methods() {
-	ClassDB::bind_method(D_METHOD("register_server", "name", "create_callback"), &PhysicsServer3DManager::register_server);
-	ClassDB::bind_method(D_METHOD("set_default_server", "name", "priority"), &PhysicsServer3DManager::set_default_server);
-}
-
-PhysicsServer3DManager *PhysicsServer3DManager::get_singleton() {
-	return singleton;
-}
-
-void PhysicsServer3DManager::register_server(const String &p_name, const Callable &p_create_callback) {
-	//ERR_FAIL_COND(!p_create_callback.is_valid());
-	ERR_FAIL_COND(find_server_id(p_name) != -1);
-	physics_servers.push_back(ClassInfo(p_name, p_create_callback));
-	on_servers_changed();
-}
-
-void PhysicsServer3DManager::set_default_server(const String &p_name, int p_priority) {
-	const int id = find_server_id(p_name);
-	ERR_FAIL_COND(id == -1); // Not found
-	if (default_server_priority < p_priority) {
-		default_server_id = id;
-		default_server_priority = p_priority;
-	}
-}
-
-int PhysicsServer3DManager::find_server_id(const String &p_name) {
-	for (int i = physics_servers.size() - 1; 0 <= i; --i) {
-		if (p_name == physics_servers[i].name) {
-			return i;
-		}
-	}
-	return -1;
-}
-
-int PhysicsServer3DManager::get_servers_count() {
-	return physics_servers.size();
-}
-
-String PhysicsServer3DManager::get_server_name(int p_id) {
-	ERR_FAIL_INDEX_V(p_id, get_servers_count(), "");
-	return physics_servers[p_id].name;
-}
-
-PhysicsServer3D *PhysicsServer3DManager::new_default_server() {
-	if (default_server_id == -1) {
-		return nullptr;
-	}
-	Variant ret;
-	Callable::CallError ce;
-	physics_servers[default_server_id].create_callback.callp(nullptr, 0, ret, ce);
-	ERR_FAIL_COND_V(ce.error != Callable::CallError::CALL_OK, nullptr);
-	return Object::cast_to<PhysicsServer3D>(ret.get_validated_object());
-}
-
-PhysicsServer3D *PhysicsServer3DManager::new_server(const String &p_name) {
-	int id = find_server_id(p_name);
-	if (id == -1) {
-		return nullptr;
-	} else {
-		Variant ret;
-		Callable::CallError ce;
-		physics_servers[id].create_callback.callp(nullptr, 0, ret, ce);
-		ERR_FAIL_COND_V(ce.error != Callable::CallError::CALL_OK, nullptr);
-		return Object::cast_to<PhysicsServer3D>(ret.get_validated_object());
-	}
-}
-
-PhysicsServer3DManager::PhysicsServer3DManager() {
-	singleton = this;
-}
-
-PhysicsServer3DManager::~PhysicsServer3DManager() {
-	singleton = nullptr;
-}
